Refactors TransactionList into single widget
